Added cleanup hook for FormRefCounter, and a function to clear a form's refs
The ref table uses the form heap manager, so it cannot wait to be
destroyed with other static objects in Components.dll.  Used the
TESDataHandler::Clear event to clean it up in advance; also handily
clears the use tables whenever the active file list is changed.

NOTE - the issues with the form heap cleanup are masked with CSE installed,
as CSE uses a quick exit that avoids normal cleanup.